requirements.txt
langchain
langchain-community 
langchainhub 
langchain-openai 
langchain-chroma 
bs4
streamlit
python-dotenv
pypdf
PyPDF2
langchain-pinecone
pinecone-client[grpc]